V1.9: RAG-Doc changes + SME control
On the chat-bot side, you can now choose to have the AI decide which topic to use when making a query, or you can manually select one or more topics (called SME's) to help answer your questions.

Additionally, the pipeline has been tweaked to enable the playground to read in Alt-Text from pptx and docx files' images, and embed those as descriptions for images instead of generating them from scratch.